+++ tinyGL/src/Engine.h
@@ -15,12 +15,12 @@ namespace tinyGL
         static string ReadFile(const string& file_path);
         float GetAspectRatio() const {return window_aspect_ratio;}
         
-        void  SetWidthHeight(float width, float height);
-        float GetWindowWidth() const {return window_width;}
-        float GetWindowHeight() const {return window_height;}
+        void  SetWidthHeight(int width, int height);
+        int GetWindowWidth() const {return window_width;}
+        int GetWindowHeight() const {return window_height;}
     private:
-        float window_width = 1024.f;
-        float window_height = 768.f;
+        int window_width = 1024;
+        int window_height = 768;
         float window_aspect_ratio = 1.f;
         GLFWwindow* render_window = nullptr;
     };
